refactor(Problem3): 클래스 분리
