+++ src/main/java/onboarding/problem3/Constant.java
@@ -1,0 +1,7 @@
+package onboarding.problem3;
+
+public class Constant {
+    static final int EXCEPTION = -1;
+    static final int INPUT_MIN = 1;
+    static final int INPUT_MAX = 10000;
+}
